+++ src/Translations_Updater/API.php
@@ -22,7 +22,7 @@ if ( ! defined( 'WPINC' ) ) {
  *
  * @package Fragen\Translations_Updater
  */
-class API {
+trait API {
 
 	/**
 	 * Variable for setting update transient hours.
@@ -78,7 +78,7 @@ class API {
 				break;
 			case 'gitlab':
 				$arr['repo']          = 'gitlab';
-				$arr['base_uri']      = 'https://gitlab.com/api/v3';
+				$arr['base_uri']      = 'https://gitlab.com/api/v4';
 				$arr['base_download'] = 'https://gitlab.com';
 				break;
 		}