+++ src/Translations_Updater/Language_Pack.php
@@ -22,7 +22,8 @@ if ( ! defined( 'WPINC' ) ) {
  *
  * @package Fragen\Translations_Updater
  */
-class Language_Pack extends Base {
+class Language_Pack {
+	use Base;
 
 	/**
 	 * Variable containing the plugin/theme object.
@@ -42,6 +43,7 @@ class Language_Pack extends Base {
 			return;
 		}
 
+		$this->add_headers();
 		$this->repo     = $repo;
 		$this->repo_api = $api;
 	}
@@ -74,17 +76,19 @@ class Language_Pack extends Base {
 		}
 
 		if ( 'pre_set_site_transient_update_plugins' === current_filter() ) {
-			$repos        = Singleton::get_instance( 'Plugin' )->get_plugin_configs();
+			$repos        = \Fragen\Singleton::get_instance( 'Plugin', $this )->get_plugin_configs();
 			$translations = wp_get_installed_translations( 'plugins' );
 		}
 		if ( 'pre_set_site_transient_update_themes' === current_filter() ) {
-			$repos        = Singleton::get_instance( 'Theme' )->get_theme_configs();
+			$repos        = \Fragen\Singleton::get_instance( 'Theme', $this )->get_theme_configs();
 			$translations = wp_get_installed_translations( 'themes' );
 		}
 
-		$repos = array_filter( $repos, function( $e ) {
-			return isset( $e->language_packs );
-		} );
+		$repos = array_filter(
+			$repos, function( $e ) {
+				return isset( $e->language_packs );
+			}
+		);
 
 		foreach ( $repos as $repo ) {
 			foreach ( $locales as $locale ) {