+++ CHANGES.md
@@ -1,3 +1,10 @@
+#### [unreleased]
+
+#### 2.2.0 / 2018-06-23
+* refactored for traits, now requires PHP 5.4
+* updated for new `Singleton` factory
+* updated for GitLab API v4
+
 #### 1.1.0 / 2017-10-30
 * refactored `class API` to not be dependent on `class Base`